Moved graph classes into graph namespace
