+++ src/JMS/Serializer/Graph/SerializerGraphNavigator.php
@@ -16,12 +16,14 @@
  * limitations under the License.
  */
 
-namespace JMS\Serializer;
+namespace JMS\Serializer\Graph;
 
+use JMS\Serializer\Context;
 use JMS\Serializer\EventDispatcher\EventDispatcherInterface;
 use JMS\Serializer\EventDispatcher\PreSerializeEvent;
 use JMS\Serializer\Handler\HandlerRegistryInterface;
 use JMS\Serializer\Metadata\ClassMetadata;
+use JMS\Serializer\TypeDefinition;
 use Metadata\MetadataFactoryInterface;
 
 /**
