+++ src/JMS/Serializer/Handler/ArrayCollectionHandler.php
@@ -21,7 +21,7 @@ namespace JMS\Serializer\Handler;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use JMS\Serializer\Context;
-use JMS\Serializer\GraphNavigator;
+use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\TypeDefinition;
 use JMS\Serializer\VisitorInterface;
