+++ src/JMS/Serializer/Handler/ConstraintViolationHandler.php
@@ -19,7 +19,7 @@
 namespace JMS\Serializer\Handler;
 
 use JMS\Serializer\Context;
-use JMS\Serializer\GraphNavigator;
+use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\JsonSerializationVisitor;
 use JMS\Serializer\TypeDefinition;
 use JMS\Serializer\XmlSerializationVisitor;
