+++ src/JMS/Serializer/Handler/FormErrorHandler.php
@@ -19,12 +19,11 @@
 namespace JMS\Serializer\Handler;
 
 use JMS\Serializer\GenericSerializationVisitor;
-use JMS\Serializer\GraphNavigator;
+use JMS\Serializer\Graph\GraphNavigator;
 use JMS\Serializer\JsonSerializationVisitor;
 use JMS\Serializer\TypeDefinition;
 use JMS\Serializer\XmlSerializationVisitor;
 use JMS\Serializer\YamlSerializationVisitor;
-use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Translation\TranslatorInterface;
