+++ src/JMS/Serializer/SerializationContext.php
@@ -19,6 +19,7 @@
 namespace JMS\Serializer;
 
 use JMS\Serializer\Exception\RuntimeException;
+use JMS\Serializer\Graph\GraphNavigator;
 use Metadata\MetadataFactoryInterface;
 
 class SerializationContext extends Context
