+++ tests/JMS/Serializer/Tests/Serializer/GraphNavigatorTest.php
@@ -21,8 +21,8 @@ namespace JMS\Serializer\Tests\Serializer;
 use Doctrine\Common\Annotations\AnnotationReader;
 use JMS\Serializer\Construction\UnserializeObjectConstructor;
 use JMS\Serializer\EventDispatcher\EventDispatcher;
-use JMS\Serializer\GraphNavigator;
-use JMS\Serializer\GraphNavigatorFactory;
+use JMS\Serializer\Graph\GraphNavigator;
+use JMS\Serializer\Graph\GraphNavigatorFactory;
 use JMS\Serializer\Handler\BasicHandler;
 use JMS\Serializer\Handler\HandlerRegistry;
 use JMS\Serializer\Handler\SubscribingHandlerInterface;
